requirements.txt
absl-py==0.7.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
Django==2.1
django-bootstrap4==0.0.8
django-storages==1.7.1
django-widgets-improved==1.5.0
easydict==1.9
gast==0.2.2
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-pasta==0.1.7
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.22.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.3
numpy==1.16.1
opencv-python==4.1.0.25
Pillow==6.1.0
protobuf==3.8.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
requests==2.22.0
rsa==4.0
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorlayer==1.8.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
